-bd modificada

+++ CoolSoft/Modelo/ENTIDADES/Cliente.cs
@@ -9,13 +9,13 @@ namespace CoolSoft.DATOS.ENTIDADES
     public class Cliente
     {
         public int idCliente { get; set; }
-        public int dniCuil { get; set; }
+        public long dniCuil { get; set; }
         public string nombre { get; set; }
         public string domicilio { get; set; }
         public string localidad { get; set; }
         public string provincia { get; set; }
-        public int telefono1 { get; set; }
-        public int telefono2 { get; set; }
+        public long telefono1 { get; set; }
+        public long telefono2 { get; set; }
 
 
         public Cliente() { }
